category and product relation bug fix

+++ app/src/main/java/io/github/yamin8000/cafe/db/entities/day/Day.kt
@@ -8,5 +8,5 @@ import java.time.LocalDate
 data class Day(
     val date: LocalDate,
     var lastOrderId: Int = 1,
-    @PrimaryKey(autoGenerate = true) val id: Int = 0
+    @PrimaryKey(autoGenerate = true) val id: Long = 0
 )
